Update README.md
In the ever-evolving landscape of healthcare, data-driven approaches are playing a pivotal role in predicting and managing various medical conditions. This project leverages MySQL, a powerful relational database management system, to develop a predictive model for assessing the risk of diabetes in females. To achieve this, we will harness a comprehensive diabetes dataset, which contains a wide array of health-related features and outcomes.

This endeavour underscores the profound impact of data analytics on healthcare, enabling us to take proactive steps towards a healthier future.